pyproject.toml
[project]
name = "patchpilot"
version = "2.0.0"
description = "Audit-grade CI repair automation. Multi-agent control plane with cryptographic audit trail."
readme = "README.md"
requires-python = ">=3.11"
license = { text = "Apache-2.0" }
authors = [{ name = "Lexiie" }]
keywords = [
    "ci",
    "github-actions",
    "ai-agent",
    "agentfield",
    "tokenrouter",
    "audit",
    "compliance",
]

dependencies = [
    "agentfield>=0.1.84",
    "pydantic>=2.5",
    "pyyaml>=6.0",
    "click>=8.1",
    "fastapi>=0.110",
    "uvicorn[standard]>=0.27",
    "httpx>=0.27",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=4.1",
    "ruff>=0.4",
    "mypy>=1.8",
]

[project.scripts]
patchpilot = "cli.main:main"

[project.urls]
Homepage = "https://github.com/Lexiie/PatchPilot"
Repository = "https://github.com/Lexiie/PatchPilot"
Issues = "https://github.com/Lexiie/PatchPilot/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["agents", "shared", "cli", "webhook"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true
